requirements.txt
Django==3.2.4
django-advanced-filters==1.3.0
django-allauth==0.45.0
django-braces==1.14.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-dbbackup==3.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
nepali-date==2.0.5
psycopg2==2.8.6
pytz==2021.1
simplejson==3.17.2
sqlparse==0.4.1
